+++ demo/config/shrine.rb
@@ -3,6 +3,7 @@
 
 require "./config/credentials"
 require "shrine"
+require "dry-monitor"
 
 # needed by `backgrounding` plugin
 require "./jobs/promote_job"
@@ -35,7 +36,7 @@ else
 end
 
 Shrine.plugin :sequel
-Shrine.plugin :logging
+Shrine.plugin :instrumentation, notifications: Dry::Monitor::Notifications.new(:test)
 Shrine.plugin :determine_mime_type
 Shrine.plugin :cached_attachment_data
 Shrine.plugin :restore_cached_data