+++ lib/shrine/plugins/instrumentation.rb
@@ -6,40 +6,58 @@ class Shrine
     #
     # [doc/plugins/instrumentation.md]: https://github.com/shrinerb/shrine/blob/master/doc/plugins/instrumentation.md
     module Instrumentation
-      EVENTS = %i[
-        storage_upload
-        storage_download
-        storage_exists
-        storage_delete
-        metadata
-      ].freeze
+      EVENTS = %i[upload download exists delete metadata].freeze
+
+      # We use a proc in order to be able identify listeners.
+      LOG_SUBSCRIBER = -> (event) { LogSubscriber.call(event) }
 
       def self.configure(uploader, opts = {})
-        uploader.opts[:instrumentation] ||= { log_subscriber: LogSubscriber.new }
+        uploader.opts[:instrumentation] ||= { log_subscriber: LOG_SUBSCRIBER, log_events: EVENTS }
         uploader.opts[:instrumentation].merge!(opts)
-        uploader.opts[:instrumentation][:notifications] ||= ActiveSupport::Notifications
+        uploader.opts[:instrumentation][:notifications] ||= ::ActiveSupport::Notifications
 
-        uploader.send(:attach_log_subscriber)
+        # we assign it to the top-level so that it's duplicated on subclassing
+        uploader.opts[:instrumentation_subscribers] ||= Hash.new { |h, k| h[k] = [] }
+
+        uploader.opts[:instrumentation][:log_events].each do |event_name|
+          uploader.subscribe(event_name, &uploader.opts[:instrumentation][:log_subscriber])
+        end
       end
 
       module ClassMethods
+        # Sends a `*.shrine` event.
+        #
+        #     # sends a `my_event.shrine` event
+        #     Shrine.instrument(:my_event) do
+        #       # work
+        #     end
         def instrument(event_name, payload = {}, &block)
           payload[:uploader] = self
 
           notifications.instrument("#{event_name}.shrine", payload, &block)
         end
 
-        def subscribe(event_name, &block)
-          notifications.subscribe("#{event_name}.shrine", &block)
+        # Subscribes to a `*.shrine` event. It rejects duplicate subscribers.
+        #
+        #     # subscribes to the `storage_upload.shrine` event
+        #     Shrine.subscribe(:storage_upload) do |event|
+        #       event.name #=> :storage_upload
+        #       event.payload #=> { location: "...", ... }
+        #       event[:location] #=> "..."
+        #       event.duration #=> 50 (in milliseconds)
+        #     end
+        def subscribe(event_name, &subscriber)
+          return if subscriber.nil?
+          return if subscribers[event_name].include?(subscriber)
+
+          notifications.subscribe("#{event_name}.shrine") do |event|
+            subscriber.call(event) if event[:uploader] <= self
+          end
+
+          subscribers[event_name] << subscriber
         end
 
         private
-
-        def attach_log_subscriber
-          EVENTS.each do |event_name|
-            subscribe(event_name) { |event| log_subscriber&.call(event) }
-          end
-        end
 
         def notifications
           Notifications.new(opts[:instrumentation][:notifications])
@@ -48,14 +66,19 @@ class Shrine
         def log_subscriber
           opts[:instrumentation][:log_subscriber]
         end
+
+        def subscribers
+          opts[:instrumentation_subscribers]
+        end
       end
 
       module InstanceMethods
         private
 
+        # Sends a `storage_upload.shrine` event.
         def copy(io, context)
           self.class.instrument(
-            :storage_upload,
+            :upload,
             storage: storage_key,
             location: context[:location],
             io: io,
@@ -64,6 +87,7 @@ class Shrine
           ) { super }
         end
 
+        # Sends a `metadata.shrine` event.
         def get_metadata(io, context)
           return super if io.is_a?(UploadedFile) && context[:metadata] != true || context[:metadata] == false
 
@@ -77,33 +101,37 @@ class Shrine
       end
 
       module FileMethods
+        # Sends a `storage_download.shrine` event.
         def open(**options)
           shrine_class.instrument(
-            :storage_download,
+            :download,
             storage: storage_key.to_sym,
             location: id,
             download_options: options,
           ) { super }
         end
 
+        # Sends a `storage_exists.shrine` event.
         def exists?
           shrine_class.instrument(
-            :storage_exists,
+            :exists,
             storage: storage_key.to_sym,
             location: id,
           ) { super }
         end
 
+        # Sends a `storage_delete.shrine` event.
         def delete
           shrine_class.instrument(
-            :storage_delete,
+            :delete,
             storage: storage_key.to_sym,
             location: id,
           ) { super }
         end
       end
 
-      # Abstracts away different types of notifications objects.
+      # Abstracts away different types of notifications objects
+      # (`ActiveSupport::Notifications` and `Dry::Monitor::Notifications`).
       class Notifications
         attr_reader :notifications
 
@@ -146,7 +174,8 @@ class Shrine
         end
       end
 
-      # Abstracts away different kind of event objects.
+      # Abstracts away different kind of event objects
+      # (`ActiveSupport::Notifications::Event` and `Dry::Events::Event`).
       class Event
         attr_reader :event
 
@@ -155,7 +184,7 @@ class Shrine
         end
 
         def name
-          library_send(:name).chomp(".shrine")
+          library_send(:name).chomp(".shrine").to_sym
         end
 
         def payload
@@ -199,72 +228,73 @@ class Shrine
           end
         end
       end
+
+      # Logs received events.
+      class LogSubscriber
+        # Entry point for logging.
+        def self.call(event)
+          new.public_send(:"on_#{event.name}", event)
+        end
+
+        def on_upload(event)
+          log "Upload (#{event.duration}ms) – #{format(
+            storage:        event[:storage],
+            location:       event[:location],
+            io:             event[:io].class,
+            upload_options: event[:upload_options],
+            uploader:       event[:uploader],
+          )}"
+        end
+
+        def on_download(event)
+          log "Download (#{event.duration}ms) – #{format(
+            storage:          event[:storage],
+            location:         event[:location],
+            download_options: event[:download_options],
+            uploader:         event[:uploader],
+          )}"
+        end
+
+        def on_exists(event)
+          log "Exists (#{event.duration}ms) – #{format(
+            storage:  event[:storage],
+            location: event[:location],
+            uploader: event[:uploader],
+          )}"
+        end
+
+        def on_delete(event)
+          log "Delete (#{event.duration}ms) – #{format(
+            storage:  event[:storage],
+            location: event[:location],
+            uploader: event[:uploader],
+          )}"
+        end
+
+        def on_metadata(event)
+          log "Metadata (#{event.duration}ms) – #{format(
+            storage:  event[:storage],
+            io:       event[:io].class,
+            uploader: event[:uploader],
+          )}"
+        end
+
+        private
+
+        def format(properties = {})
+          properties.inspect
+        end
+
+        def log(message)
+          logger.info(message)
+        end
+
+        def logger
+          Shrine.logger
+        end
+      end
     end
 
     register_plugin(:instrumentation, Instrumentation)
-  end
-
-  class LogSubscriber
-    # Entry point for logging.
-    def call(event)
-      public_send(:"on_#{event.name}", event)
-    end
-
-    def on_storage_upload(event)
-      log "Upload (#{event.duration}ms) – #{format(
-        storage:        event[:storage],
-        location:       event[:location],
-        io:             event[:io].class,
-        upload_options: event[:upload_options],
-        uploader:       event[:uploader],
-      )}"
-    end
-
-    def on_storage_download(event)
-      log "Download (#{event.duration}ms) – #{format(
-        storage:          event[:storage],
-        location:         event[:location],
-        download_options: event[:download_options],
-        uploader:         event[:uploader],
-      )}"
-    end
-
-    def on_storage_exists(event)
-      log "Exists (#{event.duration}ms) – #{format(
-        storage:  event[:storage],
-        location: event[:location],
-        uploader: event[:uploader],
-      )}"
-    end
-
-    def on_storage_delete(event)
-      log "Delete (#{event.duration}ms) – #{format(
-        storage:  event[:storage],
-        location: event[:location],
-        uploader: event[:uploader],
-      )}"
-    end
-
-    def on_metadata(event)
-      log "Metadata (#{event.duration}ms) – #{format(
-        storage:  event[:storage],
-        io:       event[:io].class,
-        uploader: event[:uploader],
-      )}"
-    end
-
-    private
-
-    def format(properties = {})
-      properties.inspect
-    end
-
-    def log(message)
-      logger.info(message)
-    end
-
-    def logger
-      Shrine.logger
-    end
   end
 end