+++ lib/shrine/plugins/logging.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Shrine.deprecation("The logging plugin has been deprecated in favor of instrumentation plugin. It will be removed in Shrine 3.")
+Shrine.deprecation("The logging plugin has been deprecated in favor of instrumentation plugin. It will no longer be available in Shrine 3.")
 
 require "logger"
 require "json"