+++ test/support/logging_helper.rb
@@ -2,6 +2,24 @@ require "stringio"
 
 class Minitest::HooksSpec
   def assert_logged(pattern)
+    result = nil
+    logged = capture_logged { result = yield }
+
+    assert_match pattern, logged
+
+    result
+  end
+
+  def refute_logged(pattern)
+    result = nil
+    logged = capture_logged { result = yield }
+
+    refute_match pattern, logged
+
+    result
+  end
+
+  def capture_logged
     previous_logger = Shrine.logger
     output = StringIO.new
     Shrine.logger = Logger.new(output)
@@ -9,7 +27,7 @@ class Minitest::HooksSpec
 
     yield
 
-    assert_match pattern, output.string
+    output.string
   ensure
     Shrine.logger = previous_logger
   end